Add sortable tables including custom for HMT
In order to allow has_many :through to be sortable by the join model
I needed to add some annotations to the table's data attributes so I 
could send those details through to the controller and conditionally
pass along the original item ids, typically, or the has_many :through
ids when sorting Exercises in the context of a Workout which needs to 
hold the sort detail on the exercise_workouts join table scoped to the
workout itself.

I don't like all the extra stuff I had to do though...especially having
to use an n+1 query on the sort controller to get the actual join model
ids to loop through for incrementing the sort order. I'm going to look
into a way to include the join model ids as the row_id directly so that
I don't have to do that on the back end.